target-app/requirements.txt
# Target App — Translation Agent (EC2 per-service venv)
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
google-genai==1.66.0
python-dotenv==1.2.2
